+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added [logi2dot] enable legend block.
 - Added [mdgrep] grep changelog example.
+- Added [Rename-Normalize] filter "replace-characters-to-avoid-in-filename"
 
 
 ## [3.7.0] - 2023-04-07